+++ h9-api/src/main/java/com/h9/api/model/vo/StickRewardMoneyVO.java
@@ -16,7 +16,7 @@ public class StickRewardMoneyVO {
 
    private String type;
 
-   private Integer rewardMoney;
+   private BigDecimal rewardMoney;
 
    private BigDecimal balance = new BigDecimal(0);
